+++ src/main/java/org/novasearch/jitter/core/selection/methods/CRCSEXP.java
@@ -15,16 +15,16 @@ public class CRCSEXP extends SelectionMethod {
     }
 
     @Override
-    public Map<String, Float> rank(List<Document> results) {
-        HashMap<String, Float> map = new HashMap<>();
+    public Map<String, Double> rank(List<Document> results) {
+        HashMap<String, Double> map = new HashMap<>();
         int j = 1;
         for (Document result : results) {
-            float r = getScore(j);
+            double r = getScore(j);
             String screenName = result.getScreen_name();
             if (!map.containsKey(screenName)) {
                 map.put(screenName, r);
             } else {
-                float cur = map.get(screenName);
+                double cur = map.get(screenName);
                 map.put(screenName, cur + r);
             }
             j++;
@@ -32,8 +32,8 @@ public class CRCSEXP extends SelectionMethod {
         return map;
     }
 
-    private float getScore(int j) {
-        return (float) (alpha * Math.exp(-beta * j));
+    private double getScore(int j) {
+        return alpha * Math.exp(-beta * j);
     }
 
     public float getAlpha() {
