+++ src/main/java/org/novasearch/jitter/core/selection/methods/CRCSISR.java
@@ -12,18 +12,18 @@ public class CRCSISR extends SelectionMethod {
     }
 
     @Override
-    public Map<String, Float> rank(List<Document> results) {
-        HashMap<String, Float> isr = new HashMap<>();
+    public Map<String, Double> rank(List<Document> results) {
+        HashMap<String, Double> isr = new HashMap<>();
 
         // sum isr
         int j = 1;
         for (Document result : results) {
-            float r = getScore(j);
+            double r = getScore(j);
             String screenName = result.getScreen_name();
             if (!isr.containsKey(screenName)) {
                 isr.put(screenName, r);
             } else {
-                float cur = isr.get(screenName);
+                double cur = isr.get(screenName);
                 isr.put(screenName, cur + r);
             }
             j++;
@@ -31,8 +31,8 @@ public class CRCSISR extends SelectionMethod {
         return isr;
     }
 
-    private float getScore(int j) {
-        return (float) (1.0 / (Math.pow(j, 2)));
+    private double getScore(int j) {
+        return 1.0 / Math.pow(j, 2);
     }
 
 }
