+++ src/main/java/org/novasearch/jitter/core/selection/methods/CRCSLOGISR.java
@@ -12,35 +12,35 @@ public class CRCSLOGISR extends SelectionMethod {
     }
 
     @Override
-    public Map<String, Float> rank(List<Document> results) {
-        HashMap<String, Float> isr = new HashMap<>();
+    public Map<String, Double> rank(List<Document> results) {
+        HashMap<String, Double> isr = new HashMap<>();
 
         // sum isr
         int j = 1;
         for (Document result : results) {
-            float r = getScore(j);
+            double r = getScore(j);
             String screenName = result.getScreen_name();
             if (!isr.containsKey(screenName)) {
                 isr.put(screenName, r);
             } else {
-                float cur = isr.get(screenName);
+                double cur = isr.get(screenName);
                 isr.put(screenName, cur + r);
             }
             j++;
         }
 
-        Map<String, Float> counts = getCounts(results);
+        Map<String, Double> counts = getCounts(results);
 
         // log multiplication
-        for (Map.Entry<String, Float> entry : isr.entrySet()) {
-            float count = counts.get(entry.getKey());
-            isr.put(entry.getKey(), (float) Math.log(1.0 + count) * entry.getValue());
+        for (Map.Entry<String, Double> entry : isr.entrySet()) {
+            double count = counts.get(entry.getKey());
+            isr.put(entry.getKey(), Math.log(1.0 + count) * entry.getValue());
         }
         return isr;
     }
 
-    private float getScore(int j) {
-        return (float) (1.0 / (Math.pow(j, 2)));
+    private double getScore(int j) {
+        return 1.0 / Math.pow(j, 2);
     }
 
 }
