+++ src/main/java/org/novasearch/jitter/core/selection/methods/ReDDE.java
@@ -13,15 +13,15 @@ public class ReDDE extends SelectionMethod {
 
     // TODO: implement ReDDE
     @Override
-    public Map<String, Float> rank(List<Document> results) {
-        HashMap<String, Float> map = new HashMap<>();
+    public Map<String, Double> rank(List<Document> results) {
+        HashMap<String, Double> map = new HashMap<>();
         for (Document result : results) {
             String screenName = result.getScreen_name();
             if (!map.containsKey(screenName)) {
-                map.put(screenName, 1f);
+                map.put(screenName, 1d);
             } else {
-                float cur = map.get(screenName);
-                map.put(screenName, cur + 1f);
+                double cur = map.get(screenName);
+                map.put(screenName, cur + 1d);
             }
         }
         return map;
