+++ src/main/java/org/novasearch/jitter/core/selection/methods/SelectionMethod.java
@@ -9,26 +9,26 @@ public abstract class SelectionMethod {
     protected SelectionMethod() {
     }
 
-    public Map<String, Float> getCounts(List<Document> results) {
-        Map<String, Float> counts = new HashMap<>();
+    public Map<String, Double> getCounts(List<Document> results) {
+        Map<String, Double> counts = new HashMap<>();
         for (Document result : results) {
             String screenName = result.getScreen_name();
             if (!counts.containsKey(screenName)) {
-                counts.put(screenName, 1f);
+                counts.put(screenName, 1d);
             } else {
-                float cur = counts.get(screenName);
-                counts.put(screenName, cur + 1f);
+                double cur = counts.get(screenName);
+                counts.put(screenName, cur + 1d);
             }
         }
         return counts;
     }
 
-    public Map<String, Float> getRanked(List<Document> results) {
-        Map<String, Float> map = rank(results);
+    public Map<String, Double> getRanked(List<Document> results) {
+        Map<String, Double> map = rank(results);
         return map;
     }
 
 
-    public abstract Map<String, Float> rank(List<Document> results);
+    public abstract Map<String, Double> rank(List<Document> results);
 
 }
